sdk: fix connection delay
The tonic client would always wait for the first message to be streamed
to yield a value. Since `stream_data` doesn't return anything until a
configuration is sent, this resulted in a deadlock.

This change requires the caller to pass a configuration stream so that
they can queue a configuration message before connecting.