Add pre-commit hook
**Summary**

It's useful to run the same checks that we run in the CI before commit.
We use the pre-commit-hooks flake to automatically generate the
pre-commit configuration based on the description in the nix flake.

**Test Plane**

Create a new shell with `nix develop`, change a few files around and
then commit the changes.