sink: filter argument is always a path to a json file
**Summary**

Previously, the `--filter` argument accepted a string argument that was
interpreted as the json-encoded filter. Users had to start the string
with @ to pass a path to a file.

This behaviour was confusing, on top of data rarely filters are simple
enough to be conveniente to type as cli argument.

Now the filter argument is always a path to a json file.

**Test Plan**

Check sinks can be run with the new cli.