starknet: support new mutable pending blocks
**Summary**

In the past, pending blocks were immutable (that is, once a pending
block was produced, it wouldn't change until becoming accepted).
Now, pending blocks grow while waiting to be accepted. Since the
sequencer is centralized, if a transaction is inserted in the block it
won't be removed from it. We use this assumption to check if a pending
block changed by counting the number of transactions in it.

**Test Plan**

Run the `apibara-starknet` service against a network that supports
Starknet v0.12. Then start streaming data by requesting pending blocks,
and see that you will receive multiple pending blocks in-between accepted
blocks.