starknet: fix cursor producer monotonicity
**Summary**

The pending cursor was being reset on receiving a new accepted block.
The code assumed that the new finalized cursor would be delivered soon
after that, but in practice the data stream would poll for new cursors
before that happened. This resulted in users getting accepted cursors
even tho the cursor was supposed to be finalized.

**Test Plane**

To reproduce the bug:

 - Stream data from the genesis block with a filter that doesn't match
   any data
 - Wait for a new accepted block to be ingested
 - Notice that suddenly you will receive a large number of empty
   accepted batches

After this commit, the bug is no longer visible.