starknet: add Provider method for pending blocks
**Summary**

Previously the ingestion engine would fetch pending blocks using the
same method as the accepted/finalized blocks. This method returns an
error if the block is not found, which is often the case for pending
blocks. This resulted in confusing logs for operators.

This commit adds a new method that returns `None` if there is no pending
block, resulting in much clearer logs.