sink: use a single javascript file for sink configuration
**Summary**

The developer experience of sinks was far from nice because options were
scattered across the cli, the filter json file, and the transform
script.

This PR changes the sinks to have scripts as the central component. All
options can be configured in the script. Operators can still override
some options by launching the sink with different command line options
or by setting environment variables.

This PR also introduces the `apibara` cli tool. It provides a `run`
command that, based on the sink specified in the script, runs the
correct sink executable.

**Test Plan**

New code is generally tested, we also run the indexers on a live stream
to see that they don't produce any error.