starknet: improve default logging experience
**Summary**

Improve the output with the default logging settings (RUST_LOG=info),
by showing one log at the beginning of the ingestion phase, and one at
the end.

Change the span level for the provider methods to avoid spamming the
OpenTelemetry collector when ingesting blocks with hundreds of
transactions.

**Test Plan**

Run apibara-starknet against a node and check that logs are produced for
every ingested block.