starknet: add flag to change head refresh interval
**Summary**

Previously the value was hardcoded to 3 seconds because it made sense
for Starknet (public).
With Madara, this value is too high and needs to be customized based on
the chain's speed.

This PR adds a new `--head-refresh-internal-ms` flag to customize the
interval at which the ingestion components checks for new pending
blocks or heads.