sink-webhook: send each block of data individually in raw mode
**Summary**

With the new version of the transform function taking a single block at the
time, it means that developers didn't have fine-grained control over the data
being sent to the server anymore (since values were accumulated into an array).

This commit changes that by sending each item returned by the transform as a
separate request.