sink-mongo: introduce entity mode
**Summary**

Sometimes users are interested in collecting a "snapshot" of their state and
not a list of changes. Entity mode is used to update entities in a database
by merging the old values with the changes produced by the indexer.

Entity mode can be used, for example, to track account balances over times.