ci: fix e2e tests running with podman
**Summary**

Our CI runs on Kubernetes, which means it cannot run Docker directly.
We resort to running a Podman system service (`podman system service`)
inside the test container.

The rust implementation of testcontainers is not 100% compatible with
podman out of the box, so we use a fork of the crate that fixes this
incompatibility for our use case.