sink: move status service to separate task
**Summary**
Previously, the indexer and the status service were running on the same
tokio task. This is an issue because the status service had to wait for
the DNA stream to finish producing messages before it could handle
updating the status. In practice, this resulted in annoying grpc timeout
errors for clients trying to access the sink status.

Moving the status service to its own task solves the issue.