sink: refactor persistence to store state
**Summary**
Adding support for factories requires tracking the dynamic filter built
by listening to factory events. We are going to use the same persistence
mechanism used to track the indexer's cursor to track the filter.

This commit changes the data stored in the persistence layer to be a
struct that can be extended later.