sink: implement factory mode
**Summary**
This PR introduces factory mode. Developers can export a `factory`
callback from their script to enable this mode.

The `filter` returned by the `factory` callback is used to create the
"main" data stream, whose transformed output is passed to the
integration.

This commit is not concerned with correctly handling cursor and state
persistence between restarts, or combining the main filter.